# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM Pipeline - Locação de Contas de Anúncio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Users
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID with their deals and ad accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete or deactivate a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Deals
...
  /deals/{dealId}/activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities for a deal
      security:
        - bearerAuth: []
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activities timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

    post:
      operationId: createActivity
      tags: [activities]
      summary: Add an activity to a deal
      security:
        - bearerAuth: []
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  # Ad Accounts
  /ad-accounts:
    get:
      operationId: listAdAccounts
      tags: [ad-accounts]
      summary: List ad accounts
      security:
        - bearerAuth: []
      parameters:
        - name: client_id
          in: query
          schema:
            type: integer
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ad accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccountListResponse"

    post:
      operationId: createAdAccount
      tags: [ad-accounts]
      summary: Create a new ad account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdAccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccount"
  /ad-accounts/{id}:
    get:
      operationId: getAdAccount
      tags: [ad-accounts]
      summary: Get an ad account by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccount"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAdAccount
      tags: [ad-accounts]
      summary: Update an ad account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdAccountUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdAccount"

    delete:
      operationId: deleteAdAccount
      tags: [ad-accounts]
      summary: Delete an ad account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationKey:
      type: apiKey
      in: header
      name: X-Integration-Key
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # Auth
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # Users
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, sales]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        document:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        document:
          type: string
        notes:
          type: string
        status:
          type: string
        externalId:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        document:
          type: string
        notes:
          type: string
        status:
          type: string
    ClientListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ClientDetail:
      type: object
      required: [client, deals, adAccounts]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        adAccounts:
          type: array
          items:
            $ref: "#/components/schemas/AdAccount"

    # Deals
    Deal:
      type: object
      required: [id, title, stage, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        stage:
          type: string
          enum: [lead_captado, qualificacao, proposta, negociacao, fechamento, onboarding, ativo, renovacao, encerrado]
        platform:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        niche:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        hasBlockHistory:
          type: ["boolean", "null"]
        blockDescription:
          type: ["string", "null"]
        proposalValue:
          type: ["number", "null"]
        proposalConditions:
          type: ["string", "null"]
        proposalTerm:
          type: ["string", "null"]
        acceptanceDate:
          type: ["string", "null"]
          format: date
        paymentMethod:
          type: ["string", "null"]
        termsAccepted:
          type: ["boolean", "null"]
        adAccountId:
          type: ["integer", "null"]
        adAccountCreated:
          type: ["boolean", "null"]
        adAccountLimit:
          type: ["number", "null"]
        firstCampaignActive:
          type: ["boolean", "null"]
        rentalStatus:
          type: ["string", "null"]
        contractEndDate:
          type: ["string", "null"]
          format: date
        clientName:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Activity:
      type: object
      required: [id, dealId, type, description, createdAt]
      properties:
        id:
          type: integer
        dealId:
          type: integer
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        type:
          type: string
          enum: [note, stage_change, contact]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ActivityInput:
      type: object
      required: [type, description]
      properties:
        type:
          type: string
          enum: [note, stage_change, contact]
        description:
          type: string
          minLength: 1
...
    AdAccount:
      type: object
      required: [id, platform, accountIdentifier, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        platform:
          type: string
        accountIdentifier:
          type: string
        monthlyLimit:
          type: ["number", "null"]
        status:
          type: string
          enum: [ativa, bloqueada, em_revisao, encerrada]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    AdAccountInput:
      type: object
      required: [platform, accountIdentifier, status]
      properties:
        clientId:
          type: integer
        platform:
          type: string
          minLength: 1
        accountIdentifier:
          type: string
          minLength: 1
        monthlyLimit:
          type: number
        status:
          type: string
          enum: [ativa, bloqueada, em_revisao, encerrada]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    AdAccountUpdate:
      type: object
      properties:
        clientId:
          type: integer
        platform:
          type: string
        accountIdentifier:
          type: string
        monthlyLimit:
          type: number
        status:
          type: string
          enum: [ativa, bloqueada, em_revisao, encerrada]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    AdAccountListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdAccount"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Dashboard
...