# lib/api-zod/src/generated/types/adAccountInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM Pipeline - Locação de Contas de Anúncio
 * OpenAPI spec version: 0.1.0
 */
import type { AdAccountInputStatus } from './adAccountInputStatus';

export interface AdAccountInput {
  clientId?: number;
  /** @minLength 1 */
  platform: string;
  /** @minLength 1 */
  accountIdentifier: string;
  monthlyLimit?: number;
  status: AdAccountInputStatus;
  startDate?: Date;
  endDate?: Date;
}