# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM Pipeline - Locação de Contas de Anúncio
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}